island: move NavigationService into island (Type-B: navigation domain)
Relocates the Google Nav SDK service (the biggest Type-B: 663-line service + 2 utils +
a 15-method adapter + 9 consumption sites) into island:

- src/services/NavigationService.ts + navigation/{routesApiCodec,roadNameResolver}.ts ->
  modules/island/src/services/ (git mv). Only host coupling was useSettingsStore
  (@/stores/settings -> ../stores/settings); restComms/CrustModule/nav-utils resolve
  in island as-is. The service is the impl (consumes no runtime hooks).
- NavigationHandlers (island) calls navigationService directly across all 9 sites
  instead of getRuntimeHooks().navigation; drops the now-dead getRuntimeHooks import.
- Removes the navigation hook from the RuntimeHooks seam (NavigationAdapter deleted;
  the Nav* DATA types stay — the handlers use them) and the 15-method host MantleManager
  configureRuntime binding + import.
- Exports navigationService from the island index; the host dev screen (debug.tsx)
  imports it from @mentra/island.

routesApiCodec.test.ts (CI-gated jest) stays in src/ and imports the moved util by
path, so it keeps running in CI (still in the 282).

tsc + island emit build clean; full jest 282 green (incl. routesApiCodec).
