+++ mobile/modules/island/src/runtime/config.ts
@@ -261,38 +261,9 @@ export type NavTripSnapshot = {
   speedLimitMps?: number | null
 }
 
-/**
- * Navigation adapter — surface of the host's NavigationService that the
- * runtime needs to wire `navigation_*` streams + request handlers for
- * miniapps.
- */
-export interface NavigationAdapter {
-  getState: () => "idle" | "navigating" | "rerouting" | "arrived"
-  getSnapshot: () => NavTripSnapshot | null
-  addListener: (l: (u: NavUpdate) => void) => () => void
-  addLocationListener: (l: (loc: NavLocation) => void) => () => void
-  addRouteListener: (l: (route: NavRoute) => void) => () => void
-  start: (
-    coords: {lat: number; lng: number},
-    options?: {
-      simulate?: boolean
-      speedMultiplier?: number
-      stops?: Array<{lat: number; lng: number}>
-      mode?: string
-      avoid?: {highways?: boolean; tolls?: boolean; ferries?: boolean}
-      missedTurnRerouteMeters?: number
-    },
-  ) => Promise<{ok: boolean; error?: string}>
-  stop: () => Promise<{ok: boolean; error?: string}>
-  simulateDeviation: (offsetMeters?: number) => Promise<{ok: boolean; error?: string}>
-  setWrongSidewalkOffset: (enabled: boolean) => Promise<{ok: boolean; error?: string}>
-  setSkipCrossings: (enabled: boolean) => Promise<{ok: boolean; error?: string}>
-  requestPermission: () => Promise<{ok: boolean; accepted: boolean; error?: string}>
-  // NOTE: route compute + reverse geocoding moved off this native adapter to the
-  // v2 cloud maps service (CloudRuntimeMapsAdapter). NavigationHandlers route the
-  // miniapp NAVIGATION_COMPUTE_ROUTE / NAVIGATION_REVERSE_GEOCODE requests to
-  // cloud.maps; this adapter now only covers the native turn-by-turn Nav SDK.
-}
+// Navigation moved into island (NavigationService, called directly by the runtime's
+// NavigationHandlers) — no longer a host-provided hook. The Nav* data types below stay
+// (the runtime handlers use them).
 
 /**
  * Heading adapter — compass / device heading subscription. The host's
@@ -410,9 +381,6 @@ export interface RuntimeHooks {
   /** Returns the connected glasses' status snapshot. */
   glassesStatus?: StoreAccessor<GlassesSnapshot>
   settings?: SettingsAccessor
-  /** Native Navigation SDK adapter (live turn-by-turn). Route compute + reverse
-   * geocoding moved to the v2 cloud maps adapter. */
-  navigation?: NavigationAdapter
   // Device heading / compass moved into island (HeadingService, subscribed
   // directly by the runtime) — no longer a host-provided hook.
   /** Location-tier escalation (e.g. realtime GPS when a trip is active). */
