+++ mobile/src/app/miniapps/settings/debug.tsx
@@ -16,7 +16,7 @@ import {useAppTheme} from "@/contexts/ThemeContext"
 import {useNavigationStore} from "@/stores/navigation"
 import {translate} from "@/i18n"
 import {SETTINGS, useSetting} from "@/stores/settings"
-import navigationService from "@/services/NavigationService"
+import {navigationService} from "@mentra/island"
 import ws from "@/services/WebSocketManager"
 import socketComms from "@/services/SocketComms"
 import showAlert from "@/utils/AlertUtils"
